extract ChachePolicyTests
